feat(ember): extract shared functions into utils
