+++ ember/app/ui/components/filters/dropdown/component.js
@@ -1,8 +1,8 @@
 import { action } from "@ember/object";
 import Component from "@glimmer/component";
-import { tracked } from "@glimmer/tracking";
 
 import { arrayFromString } from "mysagw/utils/query-params";
+
 export default class FiltersDropdownComponent extends Component {
   get selected() {
     const selected = arrayFromString(this.args.selected ?? "");
