+++ ember/app/ui/components/filters/identity/template.hbs
@@ -1,7 +1,6 @@
 <div ...attributes>
   <label class="uk-display-block">
     {{t "documents.filters.identity"}}
-    {{log @selected}}
     <PowerSelectMultiple
       @options={{this.identityOptions.value}}
       @selected={{this.selected}}
@@ -17,9 +16,14 @@
       {{option.label}}
     </PowerSelectMultiple>
   </label>
-  {{#if @invertable}}
+  {{#if @invertFilter}}
     <label>
-      <Input class="uk-checkbox" @type="checkbox"/>
+      <Input
+        class="uk-checkbox"
+        @type="checkbox"
+        @value={{@value.inverted}}
+        {{on "click" @invertFilter}}
+      />
       {{t "components.filters.invert"}}
     </label>
   {{/if}}
