+++ ember/app/ui/components/filters/input/template.hbs
@@ -3,9 +3,14 @@
     {{@label}}
     <input value={{@value}} class="uk-input" {{on "input" @onChange}}/>
   </label>
-  {{#if @invertable}}
+  {{#if @invertFilter}}
     <label>
-      <Input class="uk-checkbox" @type="checkbox"/>
+      <Input
+        class="uk-checkbox"
+        @type="checkbox"
+        @value={{@value.inverted}}
+        {{on "click" @invertFilter}}
+      />
       {{t "components.filters.invert"}}
     </label>
   {{/if}}
