+++ ember/app/ui/components/work-item-filters/template.hbs
@@ -44,11 +44,11 @@
 
   <FilterModal
     @setFiltersAmount={{this.setFiltersAmount}}
+    @resetFilters={{@resetFilters}}
     class="uk-width-auto"
   >
     <Filters
-      @invertable={{true}}
-      class="uk-grid-divider uk-grid-match uk-child-expand"
+      class="uk-grid-match uk-flex-around uk-child-width-1-4"
       uk-grid
       as |Filter|
     >
@@ -75,40 +75,45 @@
         @selected={{@filters.taskTypes}}
         @options={{this.taskTypes.value}}
         @onChange={{perform @updateFilter "taskTypes"}}
+        @invertFilter={{perform @invertFilter "taskTypes"}}
       />
       <Filter.Identity
         @selected={{@filter.identities}}
         @onChange={{perform @updateFilter "identities"}}
+        @invertFilter={{perform @invertFilter "identities"}}
       />
       <Filter.Input
         @label={{t "work-items.filters.documentNumber"}}
         @value={{@filters.documentNumber}}
         @onChange={{perform @updateFilter "documentNumber"}}
+        @invertFilter={{perform @invertFilter "documentNumber"}}
       />
       <Filter.Input
         @label={{t "work-items.filters.answer"}}
         @value={{@filters.answerSearch}}
         @onChange={{perform @updateFilter "answerSearch"}}
+        @invertFilter={{perform @invertFilter "answerSearch"}}
       />
       <Filter.Dropdown
         @label={{t "work-items.filters.forms"}}
         @selected={{@filters.forms}}
         @options={{this.forms.value}}
         @onChange={{perform @updateFilter "forms"}}
+        @invertFilter={{perform @invertFilter "forms"}}
       />
       <Filter.Dropdown
         @label={{t "work-items.filters.expertAssociations"}}
         @selected={{@filters.expertAssociations}}
         @options={{this.expertAssociations}}
         @onChange={{perform @updateFilter "expertAssociations"}}
-        @loading={{this.questionOptions.isLoading}}
+        @invertFilter={{perform @invertFilter "expertAssociations"}}
       />
       <Filter.Dropdown
         @label={{t "work-items.filters.distributionPlans"}}
         @selected={{@filters.distributionPlans}}
         @options={{this.distributionPlans}}
         @onChange={{perform @updateFilter "distributionPlans"}}
-        @loading={{this.questionOptions.isLoading}}
+        @invertFilter={{perform @invertFilter "distributionPlans"}}
       />
       <Filter.Dropdown
         @label={{t "work-items.filters.sections"}}
@@ -116,6 +121,7 @@
         @options={{this.sections}}
         @onChange={{perform @updateFilter "sections"}}
         @loading={{this.questionOptions.isLoading}}
+        @invertFilter={{perform @invertFilter "sections"}}
       />
     </Filters>
   </FilterModal>
