+++ ember/translations/components/fr.yaml
@@ -29,8 +29,8 @@ components:
     error: "Une erreur s'est produite lors de la validation du document."
 
   filters:
-    invert: "Invert filter"
-    reset: "Reset filters"
+    invert: "Filtre inversé"
+    reset: "Réinitialisation"
     dropdown:
       loading: "Chargement..."
       empty: "Aucune entrée trouvée"
